Update configuration and move site to Netlify
The site has been online for a few months, and Google Search Console shows that most of the pages are `Discovered – currently not indexed`, which means _« The page was found by Google, but not crawled yet. Typically, Google wanted to crawl the URL but this was expected to overload the site; therefore Google rescheduled the crawl. This is why the last crawl date is empty on the report. »_ We'll see if moving it on Netlify makes a difference